MAINT: Move parsing of preconditioner options to separate file
